Backplane overhaul to enable centrizied of configuration and conversion between trapi versions
